:white_check_mark: Add tests for InterviewerSlotService.deleteSlotsByUser;

+++ src/main/java/com/intellias/intellistart/interviewplanning/model/interviewerslot/InterviewerSlotService.java
@@ -64,7 +64,7 @@ public class InterviewerSlotService {
         interviewerSlotRepository.findInterviewerSlotsByUser(user);
 
     for (InterviewerSlot interviewerSlot : interviewerSlots) {
-      if (!interviewerSlot.getBookings().isEmpty()) {
+      if (interviewerSlot.getBookings() != null && !interviewerSlot.getBookings().isEmpty()) {
         bookingService.deleteBookings(interviewerSlot.getBookings());
       }
     }
